refactor: centralize backend utils
