+++ functions/api/logout.js
@@ -1,4 +1,4 @@
-import { getDb, saveDb } from '../_utils.js';
+import { getDb, saveDb } from '../../src/utils.js';
 
 export const onRequestPost = async ({ request, env }) => {
   const db = await getDb(env);
